fix: resolve live build tags and template parsing
Extract `ParseFSRecursive` into a shared `template_utils.go` file so it
can be utilized universally, independent of the `!live` build tag used in
`data_embedded.go`. In `data_live.go`, explicitly probe for
`mainPage.gohtml` to correctly resolve the application templates path
instead of incorrectly mounting `cmd/gobookmarks/templates` during
directory traversal or testing.
